+++ internal/terminal.go
@@ -1,4 +1,4 @@
-package output
+package internal
 
 import (
 	"fmt"
