+++ pkg/jato/connector.go
@@ -1,0 +1,12 @@
+package jato
+
+type Connector interface {
+	Auth()
+	Connect()
+}
+
+type Jato struct {
+	UserCredentials
+	Devices
+	CommandExpect
+}
