+++ pkg/jato/device.go
@@ -1,11 +1,9 @@
-package device
+package jato
 
 import (
 	"encoding/json"
 	"io/ioutil"
 	"log"
-
-	"github.com/automatico/jato/credentials"
 )
 
 // Device represents a managed device
@@ -39,4 +37,4 @@ func LoadDevices(fileName string) Devices {
 	return data
 }
 
-func (d Device) Auth(c credentials.UserCredentials) {}
+func (d Device) Auth(c UserCredentials) {}
