+++ pkg/jato/expecter.go
@@ -1,4 +1,4 @@
-package expecter
+package jato
 
 import (
 	"encoding/json"
